requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
torch
langchain
PyMuPDF
accelerate
requests
